backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.2,<2.0.0
supabase>=0.5.8,<0.6.0
python-dotenv>=0.19.0,<0.20.0
openai>=0.27.0,<0.28.0
httpx>=0.23.0,<0.24.0